# vityaz-main/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
        "/api/users/info": {
            "patch": {
                "operationId": "UsersController_updateUserInformation",
                "summary": "Обновление информации о текущем пользователе",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer access token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserDto"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Успешно обновлена информация о пользователе"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "UnauthorizedException": {
                                        "description": "Пользователь поставил некорректный токен доступа",
                                        "value": {
                                            "statusCode": 401,
                                            "message": "Invalid access token",
                                            "error": "Unauthorized"
                                        }
                                    }
                                },
                                "schema": {
                                    "type": "object",
                                    "description": "Пользователь поставил некорректный токен доступа",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid access token"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "required": [
                                        "statusCode",
                                        "message"
                                    ]
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "tags": [
                    "Пользователи"
                ]
            },
            "get": {
                "operationId": "UsersController_getUserInformation",
                "summary": "Получение информации о текущем пользователе",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer access token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно получена информация о пользователе",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDto"
                                }
                            }
                        }
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "UnauthorizedException": {
                                        "description": "Пользователь поставил некорректный токен доступа",
                                        "value": {
                                            "statusCode": 401,
                                            "message": "Invalid access token",
                                            "error": "Unauthorized"
                                        }
                                    }
                                },
                                "schema": {
                                    "type": "object",
                                    "description": "Пользователь поставил некорректный токен доступа",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid access token"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "required": [
                                        "statusCode",
                                        "message"
                                    ]
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "tags": [
                    "Пользователи"
                ]
            }
        },
...
        "/api/support": {
            "post": {
                "operationId": "SupportController_sendSupportRequest",
                "summary": "Отправка заявки на поддержку",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "description": "Данные заявки",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SupportRequestDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Заявка успешно отправлена"
                    }
                },
                "tags": [
                    "Поддержка"
                ]
            }
        },
        "/api/feedback": {
            "post": {
                "operationId": "SupportController_sendFeedback",
                "summary": "Отправка отзыва",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "description": "Данные отзыва",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SupportRequestDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Отзыв успешно отправлена"
                    }
                },
                "tags": [
                    "Отзывы"
                ]
            }
        },
...
    },
    "info": {
        "title": "Магазин «Витязь»",
        "description": "",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
            "UpdateUserDto": {
                "type": "object",
                "properties": {
                    "firstname": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                }
            },
            "UserDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "readOnly": true
                    },
                    "firstname": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "firstname",
                    "lastname",
                    "email"
                ]
            },
...
            "SupportRequestDto": {
                "type": "object",
                "properties": {
                    "firstname": {
                        "type": "string",
                        "description": "Имя",
                        "example": "Иван",
                        "writeOnly": true
                    },
                    "lastname": {
                        "type": "string",
                        "description": "Фамилия",
                        "example": "Иванов",
                        "writeOnly": true
                    },
                    "email": {
                        "type": "string",
                        "description": "Адрес электронной почты пользователя",
                        "example": "smoothbronx@xenofium.com",
                        "writeOnly": true
                    },
                    "message": {
                        "type": "string",
                        "description": "Сообщение",
                        "example": "Помогите!!..",
                        "writeOnly": true
                    }
                },
                "required": [
                    "firstname",
                    "lastname",
                    "email",
                    "message"
                ]
            },
...
        }
    }
}